# PERF: reduce validation overhead in `Latitude._validate_angles`