--- a/astropy/coordinates/angles/core.py
+++ b/astropy/coordinates/angles/core.py
@@ -621,7 +621,7 @@ class Latitude(Angle):
             # Otherwise, e.g., np.array(np.pi/2, 'f4') > np.pi/2 will yield True.
             angles_view = angles_view[np.newaxis]
 
-        invalid_angles = np.any(angles_view < -limit) or np.any(angles_view > limit)
+        invalid_angles = np.any(np.abs(angles_view) > limit)
         if invalid_angles:
             raise ValueError(
                 "Latitude angle(s) must be within -90 deg <= angle <= 90 deg, "