### Description
Builds on top of #16088, also helps mitigating #13479
Specifically, this cuts about another 25 to 40% overhead on `Latitude` instantiation.

<!-- Optional opt-out -->

- [ ] By checking this box, the PR author has requested that maintainers do **NOT** use the "Squash and Merge" button. Maintainers should respect this when possible; however, the final decision is at the discretion of the maintainer that merges the PR.