@neutrinoceros 2024-02-23
Although this is technically independent of https://github.com/astropy/astropy/pull/16088, the gain is only significant *as compared to* it. I could rebase on main and open for review if that's desired. Otherwise it'll just wait for #16088 to go in first.