@pllim 2024-03-18
Because I wasn't 100% sure that all `Performance` PRs would need a benchmark. Maybe sometimes you just want to update a doc related to performance.

Even if we used Performance label, you would have to re-label or push anyway to trigger it.