@neutrinoceros 2024-03-18
I think this is working
```
| Change   | Before [7ddf5ee8]    | After [31b31463]    |   Ratio | Benchmark (Parameter)     |
|----------|----------------------|---------------------|---------|---------------------------|
| -        | 18.4±0.06μs          | 14.0±0.2μs          |    0.76 | coordinates.time_latitude |
SOME BENCHMARKS HAVE CHANGED SIGNIFICANTLY.
PERFORMANCE INCREASED.
```