@neutrinoceros 2024-03-19
I note that the benchmark actually failed on the last run:
```
 | Change   | Before [7ddf5ee8]    | After [39fbd570]    |   Ratio | Benchmark (Parameter)                                  |
|----------|----------------------|---------------------|---------|--------------------------------------------------------|
| +        | 6.54±2ms             | 10.4±2ms            |    1.59 | modeling.fitting.time_combined_gauss_1d_SLSQPLSQFitter |
| +        | 32.7±0.2μs           | 45.8±2μs            |    1.4  | io_ascii.fixedwidth.FixedWidthSuite.time_splitter      |

SOME BENCHMARKS HAVE CHANGED SIGNIFICANTLY.
Error: PERFORMANCE DECREASED.
```
However, these seem hardly related. @pllim @MridulS, are we looking at some flukes (as is apparently the case in #16135) ?